# Conflicts:
#	src/sample/Game.java

+++ src/sample/serializehelp.java
@@ -3,6 +3,8 @@ package sample;
 import java.io.Serializable;
 
 public class serializehelp implements Serializable {
+    //String saveAs;
+
     double ballX;
     double ballY;
     int ballcolor;
@@ -17,10 +19,11 @@ public class serializehelp implements Serializable {
     double ObstaclenowX;
     double ObstaclenowY;
 
-    double presentscore;
+    int presentscore;
 
     serializehelp(double ballX,double ballY,int ballcolor,double starX,double starY,double colorswitcherX,double colorswitcherY,int ObstacleNumber,
-                  double ObstaclenowX,double ObstaclenowY,double presentscore){
+                  double ObstaclenowX,double ObstaclenowY,int presentscore){
+        //this.saveAs=saveAs;
         this.ballX=ballX;
         this.ballY=ballY;
         this.ballcolor=ballcolor;
